+++ app/src/main/java/com/tp/programareamea/firebase/User.kt
@@ -7,7 +7,7 @@ data class User(
     var uid: String="",
     var name: String="",
     var email: String="",
-    var userType: String="",
+    var isAdmin: Boolean= false,
     @get:Exclude var isNew: Boolean = true,
     @get:Exclude var isCreated: Boolean = true,
     @get:Exclude var isAuthenticated:Boolean = false
