+++ app/src/main/res/navigation/auth_navigator.xml
@@ -1,5 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto" android:id="@+id/auth_navigator">
-
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/auth_navigator"
+    app:startDestination="@id/googleAuthFragment2">
+    <fragment
+        android:id="@+id/googleAuthFragment2"
+        android:name="com.tp.programareamea.login.GoogleAuthFragment"
+        android:label="GoogleAuthFragment" />
 </navigation>
